IRB
The text assumes that all the analyses are for publication and thus must be approved from IRB. However, many hospitals, especially in Korea, the ATLAS is one of main operational systems  such as ERR, not just for study for research.
Thus the users in the hospital do not required to receive or even get exampted for IRB. Basically, study over the OHDSI network is based on anonymized way of research using de-identified data. We received formal decision from our IRB that they will not care about the use of CDM and ATLS for daily analysis, but if researcheres want to publish the result, then they have to get exemption before publish. 
I am afraid if that decisive ward like "MUST" in the text can affect current favorable environment on this OHDSI network study in some counties into stiff and rigid ways back.